+++ TNT.Cryptography/Tests/Tests.csproj
@@ -62,7 +62,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="CertificateTests.cs" />
-    <Compile Include="RijndaelTests.cs" />
+    <Compile Include="SymmetricTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AsymmetricTests.cs" />
   </ItemGroup>
